Finish

+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: alopez-t <alopez-t@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/03 10:59:53 by alopez-t          #+#    #+#             */
-/*   Updated: 2026/06/03 10:59:54 by alopez-t         ###   ########.fr       */
+/*   Updated: 2026/06/04 11:37:52 by alopez-t         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
@@ -14,7 +14,7 @@
 char	*ft_strchr(const char *s, int c)
 {
 	unsigned char	letter;
-	int				i;
+	size_t			i;
 
 	i = 0;
 	letter = (unsigned char) c;
