+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: alopez-t <alopez-t@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/03 11:00:56 by alopez-t          #+#    #+#             */
-/*   Updated: 2026/06/03 11:00:57 by alopez-t         ###   ########.fr       */
+/*   Updated: 2026/06/04 11:38:35 by alopez-t         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
@@ -14,7 +14,7 @@
 char	*ft_strrchr(const char *s, int c)
 {
 	unsigned char	letter;
-	int				last_occurrence;
+	size_t			last_occurrence;
 
 	last_occurrence = ft_strlen(s);
 	letter = (unsigned char) c;
